pyproject.toml
[project]
name = "events"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.40.44",
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-ipware>=7.0.1",
    "django-phonenumber-field>=8.2.0",
    "django-ratelimit>=4.1",
    "gunicorn>=23.0.0",
    "phonenumbers>=9.0.15",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "qrcode>=8.2",
    "redis>=6.4.0",
    "whitenoise>=6.11.0",
]
